% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HolidayPalettes.R
\docType{data}
\name{Christmas}
\alias{Christmas}
\title{Christmas holiday palette}
\format{
An object of class \code{HolidayPalette} of length 1.
}
\usage{
Christmas
}
\description{
A discrete/continuous color palette for Christmas-themed plots.
}
\keyword{datasets}
