% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday_countdown.R
\name{holiday_countdown}
\alias{holiday_countdown}
\title{holiday_countdown function}
\usage{
holiday_countdown(holiday)
}
\arguments{
\item{holiday}{Choose a holiday and make sure to put it in quotes: New Years Day, Valentine, St Patricks, Juneteenth, Halloween, OR Christmas.}
}
\value{
A string message telling you how many days are left until the chosen holiday!
}
\description{
this function tells you how many days left until the holiday you inputted.
}
\examples{
holiday_countdown("Christmas")
}
