% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday_df.R
\name{holiday_df}
\alias{holiday_df}
\title{holiday_df function}
\usage{
holiday_df()
}
\value{
the function returns a dataframe which is always updated with six holidays and the number of days until them, respectively
}
\description{
this function creates a dataframe with two columns: holiday and countdown. The columns contain six rows, which state the holiday name and how many until that holiday.
}
\examples{
holiday_df()
}
