% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_holiday.R
\name{scale_color_holiday}
\alias{scale_color_holiday}
\title{scale_color_holiday()}
\usage{
scale_color_holiday(palette, ...)
}
\arguments{
\item{palette}{The only available palette at the moment is "Christmas". It has 10 colors.}

\item{...}{compatible with ggplot2, this mapping accepts any other arguments provided by ggplot2's scale_color_ and scale_fill_}
}
\value{
Applies the holiday color palette to your visualization's DISCRETE variables.
}
\description{
This function applies a holiday color palette to your visualization's DISCRETE variables.
}
\examples{
ggplot(toymaker, aes(x = job_type, y = age, color = job_type)) + geom_point(size = 3) + scale_color_holiday(Christmas)
}
