% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_holiday_c.R
\name{scale_color_holiday_c}
\alias{scale_color_holiday_c}
\title{scale_color_holiday_c()}
\usage{
scale_color_holiday_c(palette, ...)
}
\arguments{
\item{palette}{The only available palette at the moment is "Christmas".}

\item{...}{compatible with ggplot2, this mapping accepts any other arguments provided by ggplot2's scale_color_ and scale_fill_}
}
\value{
Applies the holiday color palette to your visualization's CONTINUOUS variables.
}
\description{
This function applies a holiday color palette to your visualization's CONTINUOUS variables.
}
\examples{
ggplot2::ggplot(toymaker, ggplot2::aes(x = job_type, y = age, color = age)) + ggplot2::geom_point(size = 4, alpha = .1) + scale_color_holiday_c(Christmas)
}
