% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_holiday.R
\name{scale_fill_holiday}
\alias{scale_fill_holiday}
\title{scale_fill_holiday()}
\usage{
scale_fill_holiday(palette, ...)
}
\arguments{
\item{palette}{The only available palette at the moment is "Christmas". It has 10 colors.}

\item{...}{compatible with ggplot2, this mapping accepts any other arguments provided by ggplot2's scale_color_ and scale_fill_}
}
\value{
Applies the holiday color palette to your visualization's DISCRETE variables.
}
\description{
This function applies a holiday color palette to your visualization's DISCRETE variables.
}
\examples{
ggplot2::ggplot(toymaker, ggplot2::aes(sex, fill = sex)) + ggplot2::geom_bar() + scale_fill_holiday(Christmas)
}
