% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday_df.R
\docType{data}
\name{holiday_df}
\alias{holiday_df}
\title{Holiday Countdowns}
\format{
\subsection{\code{holiday_df}}{

A data frame with 6 rows and 2 columns:
\describe{
\item{holiday}{The name of the holiday.}
\item{countdown}{The number of days until the holiday.}
}
}
}
\source{
Created by the package author, Kaylen Ton.
}
\usage{
holiday_df
}
\description{
This dataframe consists of six holidays: New Years Day, Valentine, St Patricks, Juneteenth, Halloween, and Christmas.
It contains a column which records the number of days until the holiday, updated in real time.
}
\keyword{datasets}
