% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday_df.R
\name{holiday_df}
\alias{holiday_df}
\title{holiday_df function}
\usage{
holiday_df()
}
\value{
a dataframe which is always updated with six holidays and the number of days until them, respectively
}
\description{
holiday_df function
}
\examples{
holiday_df()
}
