% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_holiday_countdowns.R
\name{plot_holiday_countdowns}
\alias{plot_holiday_countdowns}
\title{plot_holiday_countdowns function}
\usage{
plot_holiday_countdowns()
}
\value{
a bar chart in ascending order, visualizing six holidays and how far they are in days
}
\description{
This function creates a simple bar chart visualization, ascending the holidays in the order they are closet to further from today!
}
\examples{
plot_holiday_countdowns()
}
